Separated both crawlers
As per toddy's request. OFAScraper now belongs to Dao
