+++ scripts/browserCrawler/OFAScraper.py
@@ -1,3 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+Created on 01/27/2019
+NSC - AD440 CLOUD PRACTICIUM
+@author: Michael Leon
+
+Changed ownership on 02/09/2019
+@author: Dao Nguyen
+
+"""
+
 import urllib.request
 import re
 import os
@@ -10,26 +21,83 @@ from bs4 import BeautifulSoup
 import time
 from dateutil.parser import parse
 from datetime import datetime
+import boto3
 
+
+# This script scrapes a website and pulls specific data.
 FOUND_LIST = []
+QUEUE = []
 OUTPUT = {}
+DATA = {}
 SOUP = []
+OFA = "https://outdoorsforall.org/events-news/calendar/"
+#s3 = boto3.resource('s3', aws_access_key_id=AWS_ACCESS_KEY_ID, aws_secret_access_key=AWS_SECRET_ACCESS_KEY)
+
+def ofa_crawl(url):
+    print(url)
+    global QUEUE
+    global FOUND_LIST
+    global SOUP
+    jsQueue = []
+    options = webdriver.ChromeOptions()
+    options.add_argument('headless')
+    options.add_argument("--log-level=3")
+    driver = webdriver.Chrome(chrome_options=options)
+    driver.get(url)
+    soup = BeautifulSoup(driver.page_source, "html.parser")
+    pages = 3
+    #Grabs all links on calendar
+    while pages <= 3:
+        for row in soup.find_all("div"):
+            if row.get("onclick"):
+                jsQueue.append(row.get("class")[0])
+        #driver.find_element_by_xpath("//*[@id='main_cal']/tbody/tr/td/table/tbody/tr[1]/td[3]/a").click()
+        pages += 1
+        x = driver.find_elements_by_class_name(jsQueue[0])
+    count = 0
+    # Click all found elements to open page and grab the URL
+    for row in x:
+        row.click()
+        driver.switch_to.window(driver.window_handles[1])
+        if driver.current_url not in FOUND_LIST:
+            print()
+            print("Scraping for a new event...")
+            #QUEUE.append(driver.current_url)
+            FOUND_LIST.append(driver.current_url)
+            print("Links remaining - " + str(len(jsQueue)))
+            jsQueue.pop(0)
+            current_url = driver.current_url
+            current_soup = BeautifulSoup(driver.page_source, "html.parser")
+            for linebreak in current_soup.find_all('br'):
+                linebreak.extract()
+            #Calls OFAScraper module to populate a dictionary object to add to the output
+            OUTPUT[current_url] = open_link(current_soup, current_url)
+            #SOUP.append(BeautifulSoup(driver.page_source, "html.parser"))
+            driver.switch_to.window(driver.window_handles[0])
+            print("Scraping reached end of page")
+        else:
+            driver.switch_to.window(driver.window_handles[0])
+        #Count is used for test purposes only.
+        #count += 1
+        if count == 3:
+            break
+    driver.quit()
 
 def open_link(current_soup, current_url):
-    OUTPUT["URL"] = current_url
+    DATA["URL"] = current_url
     find_title(current_soup)
     find_date(current_soup)  
-    return OUTPUT   
+    return DATA   
 
 def find_title(soup):
-    global OUTPUT
+    global DATA
     if soup.find(class_="header-theme"):
         title = soup.find(class_="header-theme").text
-        OUTPUT["Title"] = title
+        DATA["Title"] = title
         find_description(soup)
 
 def find_description(soup):
-    global OUTPUT
+    global DATA
     desc = soup.find("span", attrs={"class": "event-desc-theme"})
     #print(desc)
     p_desc = ""
@@ -50,24 +118,24 @@ def find_description(soup):
         except:
             pass
     if time != "":
-        OUTPUT["Time"] = time
+        DATA["Time"] = time
     else:
-        OUTPUT["Time"] = "Unknown"
+        DATA["Time"] = "Unknown"
     for row in desc.findAll(text=True, recursive=False):
         if(len(row) > 1):
             loc = re.sub("\r\n", "", row)
             find_location(loc)
             break
-    OUTPUT["Description"] = p_desc
+    DATA["Description"] = p_desc
 
 def find_date(soup):
-    global OUTPUT
+    global DATA
     header_re = re.compile('.*header.*')
     for row in soup.findAll(attrs={"class": header_re}):
         try:
             for val in row:
                 if parse(val):
-                    OUTPUT["Date"] = str(parse(val))
+                    DATA["Date"] = str(parse(val))
                     break
         except Exception as a:
             #print(a)
@@ -79,6 +147,39 @@ def find_date(soup):
             pass
 
 def find_location(location):
-    global OUTPUT
-    OUTPUT["Location"] = location
+    global DATA
+    DATA["Location"] = location
     #print(OUTPUT)W
+
+def create_json():
+    print("creating json...")
+    with open('OFA_event_data.json', 'w') as outfile:
+        json.dump(OUTPUT, outfile)
+    #s3.Object('mjleontest', 'browser_event_data.json').put(Body=open('browser_event_data.json', 'rb'))
+
+def main():
+    start_time = time.time()
+    print("Crawler Started.")
+    count = 0
+    while count != 5:
+        try:
+            print("OFA Crawl Started.")
+            #open_url(EVENT_BRITE)
+            ofa_crawl(OFA)
+            print("OFA Crawl Completed.")
+            break
+        except Exception as e:
+            print("Error gathering URL data, " + str(e))
+            if str(e) == "list index out of range":
+                count += 1
+                print("Retrying selenium...")
+            else:
+                break
+    create_json()
+    elapsed_time = time.time() - start_time
+    print("Crawler Ended.")
+    print(elapsed_time)
+    
+if __name__ == '__main__':   
+    main()
+
